packages/ml/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
psycopg2-binary>=2.9.0
joblib>=1.3.0
